refactor(settings): consolidate integration settings for various plugins
- Removed legacy integration settings files for Awesome Support, BuddyPress, Contact Form 7, Easy Digital Downloads, Gravity Forms, Job Manager, Quform, Ultimate Member, and WooCommerce.
- Introduced new integration settings files that enhance structure and maintainability, ensuring consistent handling of SMS notifications across different plugins.
- Updated the settings to include dynamic sections and fields, improving user experience and integration capabilities.
