+++ src/Settings/Groups/Integrations/AwesomeSupportSettings.php
@@ -4,6 +4,8 @@ namespace WP_SMS\Settings\Groups\Integrations;
 
 use WP_SMS\Settings\Abstracts\AbstractSettingGroup;
 use WP_SMS\Settings\Field;
+use WP_SMS\Settings\Section;
+use WP_SMS\Settings\LucideIcons;
 use WP_SMS\Notification\NotificationFactory;
 
 class AwesomeSupportSettings extends AbstractSettingGroup
@@ -15,133 +17,140 @@ class AwesomeSupportSettings extends AbstractSettingGroup
 
     public function getLabel(): string
     {
-        return 'Awesome Support Integration Settings';
+        return __('Awesome Support', 'wp-sms');
     }
 
-    public function isAvailable(): bool
+    public function getIcon(): string
     {
-        return class_exists('Awesome_Support');
+        return LucideIcons::HELP_CIRCLE;
     }
 
-    public function getFields(): array
+    public function getSections(): array
     {
-        if (! $this->isAvailable()) {
+        if (!class_exists('Awesome_Support')) {
             return [
-                new Field([
-                    'key'         => 'as_notify_new_ticket',
-                    'type'        => 'notice',
-                    'label'       => 'Not active',
-                    'description' => 'Awesome Support plugin should be installed to show the options.',
-                    'group_label' => 'Awesome Support',
+                new Section([
+                    'id' => 'awesome_support_not_active',
+                    'title' => __('Awesome Support Integration', 'wp-sms'),
+                    'subtitle' => __('Configure SMS notifications for support tickets', 'wp-sms'),
+                    'fields' => [
+                        new Field([
+                            'key' => 'awesome_support_not_active_notice',
+                            'label' => __('Not active', 'wp-sms'),
+                            'type' => 'notice',
+                            'description' => __('Awesome Support plugin should be installed to show the options.', 'wp-sms')
+                        ])
+                    ]
                 ])
             ];
         }
 
         return [
-            new Field([
-                'key'         => 'as_notify_new_ticket',
-                'type'        => 'header',
-                'label'       => 'Notify for new ticket',
-                'group_label' => 'Awesome Support',
+            new Section([
+                'id' => 'new_ticket_notification',
+                'title' => __('Notify for new ticket', 'wp-sms'),
+                'subtitle' => __('Configure SMS notifications for new support ticket submissions', 'wp-sms'),
+                'fields' => [
+                    new Field([
+                        'key' => 'as_notify_open_ticket_status',
+                        'label' => __('Send SMS', 'wp-sms'),
+                        'type' => 'checkbox',
+                        'description' => __('Send SMS to admin when the user opened a new ticket.', 'wp-sms')
+                    ]),
+                    new Field([
+                        'key' => 'as_notify_open_ticket_message',
+                        'label' => __('Message body', 'wp-sms'),
+                        'type' => 'textarea',
+                        'description' => __('Enter the contents of the SMS message.', 'wp-sms') . '<br>' . NotificationFactory::getAwesomeSupportTicket()->printVariables()
+                    ]),
+                ]
             ]),
-            new Field([
-                'key'         => 'as_notify_open_ticket_status',
-                'type'        => 'checkbox',
-                'label'       => 'Send SMS (New Ticket)',
-                'description' => 'Send SMS to admin when a user opens a new ticket',
-                'group_label' => 'Awesome Support',
+            new Section([
+                'id' => 'admin_reply_notification',
+                'title' => __('Notify admin for get reply', 'wp-sms'),
+                'subtitle' => __('Configure SMS notifications for admin when users reply to tickets', 'wp-sms'),
+                'fields' => [
+                    new Field([
+                        'key' => 'as_notify_admin_reply_ticket_status',
+                        'label' => __('Send SMS', 'wp-sms'),
+                        'type' => 'checkbox',
+                        'description' => __('Send SMS to admin when the user replied the ticket.', 'wp-sms')
+                    ]),
+                    new Field([
+                        'key' => 'as_notify_admin_reply_ticket_message',
+                        'label' => __('Message body', 'wp-sms'),
+                        'type' => 'textarea',
+                        'description' => __('Enter the contents of the SMS message.', 'wp-sms') . '<br>' . NotificationFactory::getAwesomeSupportTicket()->printVariables()
+                    ]),
+                ]
             ]),
-            new Field([
-                'key'         => 'as_notify_open_ticket_message',
-                'type'        => 'textarea',
-                'label'       => 'Message body (New Ticket)',
-                'description' => 'SMS content. Supports Awesome Support placeholders.<br>' . NotificationFactory::getAwesomeSupportTicket()->printVariables(),
-                'group_label' => 'Awesome Support',
+            new Section([
+                'id' => 'user_reply_notification',
+                'title' => __('Notify user for get reply', 'wp-sms'),
+                'subtitle' => __('Configure SMS notifications for users when admins reply to tickets', 'wp-sms'),
+                'fields' => [
+                    new Field([
+                        'key' => 'as_notify_user_reply_ticket_status',
+                        'label' => __('Send SMS', 'wp-sms'),
+                        'type' => 'checkbox',
+                        'description' => __('Send SMS to user when the admin replied the ticket. Please make sure the "Add Mobile number field" option is enabled in the Settings > Features', 'wp-sms')
+                    ]),
+                    new Field([
+                        'key' => 'as_notify_user_reply_ticket_message',
+                        'label' => __('Message body', 'wp-sms'),
+                        'type' => 'textarea',
+                        'description' => __('Enter the contents of the SMS message.', 'wp-sms') . '<br>' . NotificationFactory::getAwesomeSupportTicket()->printVariables()
+                    ]),
+                ]
             ]),
-
-            new Field([
-                'key'         => 'as_notify_admin_reply_ticket',
-                'type'        => 'header',
-                'label'       => 'Notify admin for get reply',
-                'group_label' => 'Awesome Support',
+            new Section([
+                'id' => 'ticket_status_update_notification',
+                'title' => __('Notify user for the ticket status update', 'wp-sms'),
+                'subtitle' => __('Configure SMS notifications for users when ticket status changes', 'wp-sms'),
+                'fields' => [
+                    new Field([
+                        'key' => 'as_notify_update_ticket_status',
+                        'label' => __('Send SMS', 'wp-sms'),
+                        'type' => 'checkbox',
+                        'description' => __('Send SMS to user when the ticket status updates', 'wp-sms')
+                    ]),
+                    new Field([
+                        'key' => 'as_notify_update_ticket_message',
+                        'label' => __('Message body', 'wp-sms'),
+                        'type' => 'textarea',
+                        'description' => __('Enter the contents of the SMS message.', 'wp-sms') . '<br>' . NotificationFactory::getAwesomeSupportTicket()->printVariables()
+                    ]),
+                ]
             ]),
-            new Field([
-                'key'         => 'as_notify_admin_reply_ticket_status',
-                'type'        => 'checkbox',
-                'label'       => 'Send SMS (User Reply to Ticket)',
-                'description' => 'Send SMS to admin when the user replies to a ticket',
-                'group_label' => 'Awesome Support',
-            ]),
-            new Field([
-                'key'         => 'as_notify_admin_reply_ticket_message',
-                'type'        => 'textarea',
-                'label'       => 'Message body (User Reply to Ticket)',
-                'description' => 'SMS content. Supports Awesome Support placeholders.<br>' . NotificationFactory::getAwesomeSupportTicket()->printVariables(),
-                'group_label' => 'Awesome Support',
-            ]),
-
-            new Field([
-                'key'         => 'as_notify_user_reply_ticket',
-                'type'        => 'header',
-                'label'       => 'Notify user for get reply',
-                'group_label' => 'Awesome Support',
-            ]),
-            new Field([
-                'key'         => 'as_notify_user_reply_ticket_status',
-                'type'        => 'checkbox',
-                'label'       => 'Send SMS (Admin Reply to Ticket)',
-                'description' => 'Send SMS to user when admin replies (requires mobile field enabled)',
-                'group_label' => 'Awesome Support',
-            ]),
-            new Field([
-                'key'         => 'as_notify_user_reply_ticket_message',
-                'type'        => 'textarea',
-                'label'       => 'Message body (Admin Reply to Ticket)',
-                'description' => 'SMS content. Supports Awesome Support placeholders.<br>' . NotificationFactory::getAwesomeSupportTicket()->printVariables(),
-                'group_label' => 'Awesome Support',
-            ]),
-
-            new Field([
-                'key'         => 'as_notify_update_ticket',
-                'type'        => 'header',
-                'label'       => 'Notify user for the ticket status update',
-                'group_label' => 'Awesome Support',
-            ]),
-            new Field([
-                'key'         => 'as_notify_update_ticket_status',
-                'type'        => 'checkbox',
-                'label'       => 'Send SMS (Ticket Status Update)',
-                'description' => 'Send SMS to user when ticket status is updated',
-                'group_label' => 'Awesome Support',
-            ]),
-            new Field([
-                'key'         => 'as_notify_update_ticket_message',
-                'type'        => 'textarea',
-                'label'       => 'Message body (Ticket Status Update)',
-                'description' => 'SMS content. Supports Awesome Support placeholders.<br>' . NotificationFactory::getAwesomeSupportTicket()->printVariables(),
-                'group_label' => 'Awesome Support',
-            ]),
-
-            new Field([
-                'key'         => 'as_notify_close_ticket',
-                'type'        => 'header',
-                'label'       => 'Notify user when the ticket is closed',
-                'group_label' => 'Awesome Support',
-            ]),
-            new Field([
-                'key'         => 'as_notify_close_ticket_status',
-                'type'        => 'checkbox',
-                'label'       => 'Send SMS (Ticket Closed)',
-                'description' => 'Send SMS to user when ticket is closed',
-                'group_label' => 'Awesome Support',
-            ]),
-            new Field([
-                'key'         => 'as_notify_close_ticket_message',
-                'type'        => 'textarea',
-                'label'       => 'Message body (Ticket Closed)',
-                'description' => 'SMS content. Supports Awesome Support placeholders.<br>' . NotificationFactory::getAwesomeSupportTicket()->printVariables(),
-                'group_label' => 'Awesome Support',
+            new Section([
+                'id' => 'ticket_close_notification',
+                'title' => __('Notify user when the ticket is closed', 'wp-sms'),
+                'subtitle' => __('Configure SMS notifications for users when tickets are closed', 'wp-sms'),
+                'fields' => [
+                    new Field([
+                        'key' => 'as_notify_close_ticket_status',
+                        'label' => __('Send SMS', 'wp-sms'),
+                        'type' => 'checkbox',
+                        'description' => __('Send SMS to user when the ticket is closed', 'wp-sms')
+                    ]),
+                    new Field([
+                        'key' => 'as_notify_close_ticket_message',
+                        'label' => __('Message body', 'wp-sms'),
+                        'type' => 'textarea',
+                        'description' => __('Enter the contents of the SMS message.', 'wp-sms') . '<br>' . NotificationFactory::getAwesomeSupportTicket()->printVariables()
+                    ]),
+                ]
             ]),
         ];
     }
-}
+
+    public function getFields(): array
+    {
+        // Legacy method - return all fields from all sections for backward compatibility
+        $allFields = [];
+        foreach ($this->getSections() as $section) {
+            $allFields = array_merge($allFields, $section->getFields());
+        }
+        return $allFields;
+    }
+}
