+++ src/Settings/Groups/Integrations/ContactForm7Settings.php
@@ -4,33 +4,53 @@ namespace WP_SMS\Settings\Groups\Integrations;
 
 use WP_SMS\Settings\Abstracts\AbstractSettingGroup;
 use WP_SMS\Settings\Field;
+use WP_SMS\Settings\Section;
+use WP_SMS\Settings\LucideIcons;
 
-class ContactForm7Settings extends AbstractSettingGroup {
-    public function getName(): string {
+class ContactForm7Settings extends AbstractSettingGroup
+{
+    public function getName(): string
+    {
         return 'contact_form7';
     }
 
-    public function getLabel(): string {
-        return 'Contact Form 7 Settings';
+    public function getLabel(): string
+    {
+        return __('Contact Form 7', 'wp-sms');
     }
 
-    public function getFields(): array {
+    public function getIcon(): string
+    {
+        return LucideIcons::MAIL;
+    }
+
+    public function getSections(): array
+    {
         return [
-            new Field([
-                'key'         => 'cf7_title',
-                'type'        => 'header',
-                'label'       => 'SMS Notification Metabox',
-                'description' => 'Add SMS notification tools to all Contact Form 7 edit forms.',
-                'doc'  => '/resources/integrate-wp-sms-with-contact-form-7/',
-                'group_label' => 'Contact Form 7',
-            ]),
-            new Field([
-                'key'         => 'cf7_metabox',
-                'type'        => 'checkbox',
-                'label'       => 'Status',
-                'description' => 'Enables SMS Notification tab in Contact Form 7 form editor.',
-                'group_label' => 'Contact Form 7',
+            new Section([
+                'id' => 'sms_notification_metabox',
+                'title' => __('SMS Notification Metabox', 'wp-sms'),
+                'subtitle' => __('By this option you can add SMS notification tools in all edit forms.', 'wp-sms'),
+                'help_url' => '/resources/integrate-wp-sms-with-contact-form-7/',
+                'fields' => [
+                    new Field([
+                        'key' => 'cf7_metabox',
+                        'label' => __('Status', 'wp-sms'),
+                        'type' => 'checkbox',
+                        'description' => __('This option adds SMS Notification tab in the edit forms.', 'wp-sms')
+                    ]),
+                ]
             ]),
         ];
     }
-}
+
+    public function getFields(): array
+    {
+        // Legacy method - return all fields from all sections for backward compatibility
+        $allFields = [];
+        foreach ($this->getSections() as $section) {
+            $allFields = array_merge($allFields, $section->getFields());
+        }
+        return $allFields;
+    }
+}
