--- src/Settings/Groups/ProWordPressSettings.php
@@ -1,1 +1,0 @@
-
